+++ orders.h
@@ -5,11 +5,11 @@
 
 class Order {
     public:
-        Order(e_orderType orderType, uint64_t orderId, e_side side, uint64_t price, uint64_t initialQuantity) : c_orderType(orderType), c_orderId(orderId), c_side(side), c_price(price), c_initialQuantity(initialQuantity), c_remainingQuantity(initialQuantity) {}
+        Order(e_orderType orderType, uint64_t orderId, e_side side, double price, uint64_t initialQuantity) : c_orderType(orderType), c_orderId(orderId), c_side(side), c_price(price), c_initialQuantity(initialQuantity), c_remainingQuantity(initialQuantity) {}
         e_orderType orderType() const { return c_orderType; }
         uint64_t orderId() const { return c_orderId; }
         e_side side() const { return c_side; }
-        uint64_t price() const { return c_price; }
+        double price() const { return c_price; }
         uint64_t initialQuantity() const { return c_initialQuantity; }
         uint64_t remainingQuantity() const { return c_remainingQuantity; }
         void setRemainingQuantity(uint64_t remainingQuantity) { c_remainingQuantity = remainingQuantity; }
@@ -24,7 +24,7 @@ class Order {
         e_orderType c_orderType;
         uint64_t c_orderId;
         e_side c_side;
-        uint64_t c_price;
+        double c_price;
         uint64_t c_initialQuantity;
         uint64_t c_remainingQuantity;
 };
